+++ src/psyclone/psyir/nodes/profile_node.py
@@ -39,8 +39,15 @@
     generated by PSyclone. '''
 
 from __future__ import absolute_import, print_function
-
+from fparser.common.readfortran import FortranStringReader
+from fparser.common.sourceinfo import FortranFormat
+from fparser.two import Fortran2003
+from psyclone.psyir.nodes.codeblock import CodeBlock
 from psyclone.psyir.nodes.psy_data_node import PSyDataNode
+from psyclone.psyir.nodes.routine import Routine
+from psyclone.psyir.symbols import (ContainerSymbol, DataSymbol, TypeSymbol,
+                                    GlobalInterface, DeferredType,
+                                    INTEGER_TYPE)
 
 
 class ProfileNode(PSyDataNode):
@@ -136,3 +143,82 @@ class ProfileNode(PSyDataNode):
         '''
         raise NotImplementedError("Generation of C code is not supported "
                                   "for profiling")
+
+    # -------------------------------------------------------------------------
+    def lower_to_language_level(self):
+        '''
+        Lowers this node (and all children) to language-level PSyIR. The
+        PSyIR tree is modified in-place.
+
+        '''
+        for child in self.children:
+            child.lower_to_language_level()
+
+        if self._module_name:
+            name = self._module_name
+        else:
+            name = self.ancestor(Routine).name
+        routine_name = name # Literal(name, CHARACTER_TYPE)
+
+        symbol_table = self.scope.symbol_table
+        # Ensure that we have a container symbol for the API access
+        fortran_module = self.add_psydata_class_prefix(self.fortran_module)
+        try:
+            csym = symbol_table.lookup(fortran_module)
+        except KeyError:
+            csym = ContainerSymbol(fortran_module)
+            symbol_table.add(csym)
+
+        for symbol in self.symbols:
+            try:
+                sym = symbol_table.lookup(symbol)
+                # assert sym.interface = 
+            except KeyError:
+                sym = DataSymbol(symbol, INTEGER_TYPE,
+                                 interface=GlobalInterface(csym))
+                symbol_table.add(sym)
+
+        # The type symbol for PSyData variables
+        var_name = self.add_psydata_class_prefix("PSyDataType")
+        try:
+            type_sym = symbol_table.lookup(var_name)
+        except KeyError:
+            type_sym = TypeSymbol(var_name, DeferredType(),
+                                  interface=GlobalInterface(csym))
+
+        # Create a name for this region
+        if self._region_name:
+            name = self._region_name
+        else:
+            name = "rTODO"
+        region_name = name #Literal(name, CHARACTER_TYPE)
+
+        var_name = self.add_psydata_class_prefix("psy_data")
+
+        # Create a symbol for this PSyData region
+        # TODO This symbol needs the 'save' and 'target' attributes in Fortran
+        dsym = symbol_table.new_symbol(var_name, symbol_type=DataSymbol,
+                                       datatype=type_sym)
+
+        # PSyData end call
+        reader = FortranStringReader(
+            "CALL {0}%PostEnd".format(var_name))
+        # Tell the reader that the source is free format
+        reader.set_format(FortranFormat(True, False))
+        pecall = Fortran2003.Call_Stmt(reader)
+        end_call = CodeBlock([pecall], CodeBlock.Structure.STATEMENT)
+        self.parent.children.insert(self.position+1, end_call)
+        end_call.parent = self.parent
+
+        # PSyData start call (replaces existing PSyDataNode)
+        reader = FortranStringReader(
+            "CALL {2}%PreStart('{0}', '{1}', 0, 0)".format(
+                routine_name, region_name, var_name))
+        reader.set_format(FortranFormat(True, False))
+        pscall = Fortran2003.Call_Stmt(reader)
+        start_call = CodeBlock([pscall], CodeBlock.Structure.STATEMENT)
+
+        for child in self.children[0].pop_all_children():
+            self.parent.children.insert(self.position+1, child)
+            child.parent = self.parent
+        self.replace_with(start_call)
